+++ app/Http/Controllers/Api/RoleController.php
@@ -3,10 +3,15 @@
 namespace App\Http\Controllers\Api;
 
 use App\Role;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use App\Http\Resources\RoleResource;
 use App\Http\Controllers\Controller;
+use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
+use Illuminate\Validation\Rule;
 use Symfony\Component\HttpFoundation\Response;
 
 
@@ -21,36 +26,70 @@ class RoleController extends Controller
     }
 
     /**
+     * Get a validator for an incoming registration request.
+     *
+     * @param array $data
+     * @param Role|null $role
+     * @return \Illuminate\Contracts\Validation\Validator
+     */
+    protected function validator(array $data, Role $role = null)
+    {
+        $id = $role ? $role->id : null;
+        return Validator::make($data, [
+            'name' => [
+                'required',
+                'string',
+                'max:255',
+                Rule::unique('roles')->ignore($id)->where('group_slug', $data['group_slug'])
+            ],
+        ]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection
+     * @return JsonResponse|AnonymousResourceCollection
      */
-    public function index()
+    public function index(Request $request)
     {
-        return RoleResource::collection(Role::all());
+        return RoleResource::collection($this->role->where('group_slug', $request->group_slug)->get());
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return JsonResponse
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => ['required', 'string', 'max:255', 'unique:tenant.roles,name']
-        ]);
+        if (!Auth::user()->hasPermissionTo($request->group_slug, 'create-role')) {
+            return response()->json([
+                'message' => 'Action forbidden. You don\'t have the permission to do that',
+                'redirect' => 'home'
+            ], Response::HTTP_FORBIDDEN);
+        }
+
+        $request->name = Str::lower($request->name);
+
+        $data = $this->validator($request->all())->validate();
+
         $role = $this->role->create([
-            'name' => $request->name,
-            'display_name' => ucwords($request->name)
+            'name' => $data['name'],
+            'display_name' => Str::title($data['name']),
+            'group_slug' => $request->group_slug
         ]);
 
         if ($request->has('permissions')) {
-            $role->givePermissionTo(collect($request->permissions)->pluck('id')->toArray());
+            $role->givePermissionsToUsingSlug(
+                $request->group_slug,
+                collect($request->permissions)
+                    ->pluck('slug')
+                    ->toArray()
+            );
         }
 
-        return response([
+        return response()->json([
             'message' => 'Role created',
             'payload' => RoleResource::make($role)
         ], Response::HTTP_OK);
@@ -59,19 +98,27 @@ class RoleController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param Request $request
      * @param Role $role
-     * @return \Illuminate\Http\Response
+     * @return JsonResponse
      */
-    public function update(Request $request, Role $role) {
+    public function update(Request $request, Role $role)
+    {
+        if (!Auth::user()->hasPermissionTo($request->group_slug, 'edit-role')) {
+            return response()->json([
+                'message' => 'Action forbidden. You don\'t have the permission to do that',
+                'redirect' => 'home'
+            ], Response::HTTP_FORBIDDEN);
+        }
 
-        $request->validate([
-            'name' => ['required', 'string', 'max:255', 'unique:tenant.roles,name,'.$role->id]
-        ]);
+        $request->name = Str::lower($request->name);
+        $role->slug = null;
+
+        $data = $this->validator($request->all(), $role)->validate();
 
         $role->update([
-            'name' => $request->name,
-            'display_name' => ucwords($request->name),
+            'name' => $data['name'],
+            'display_name' => Str::title($data['name']),
             'updated_at' => now()
         ]);
 
@@ -84,7 +131,7 @@ class RoleController extends Controller
             );
         }
 
-        return response([
+        return response()->json([
             'message' => 'Role updated',
             'payload' => RoleResource::make($role)
         ], Response::HTTP_OK);
@@ -94,9 +141,17 @@ class RoleController extends Controller
      * Remove the specified resource from storage.
      *
      * @param int $id
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
-    public function destroy($id) {
+    public function destroy(Request $request, $id)
+    {
+        if (!Auth::user()->hasPermissionTo($request->group_slug, 'delete-role')) {
+            return response()->json([
+                'message' => 'Action forbidden. You don\'t have the permission to do that',
+                'redirect' => 'home'
+            ], Response::HTTP_FORBIDDEN);
+        }
+
         $this->role->destroy($id);
 
         return response()->json([
